requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
autocommand==2.2.2
blinker==1.9.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
docker==7.1.0
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.3
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gitdb==4.0.11
greenlet==3.1.1
idna==3.6
inflect==7.0.0
installer==0.7.0
ipython==8.23.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==4.3.0
jaraco.env==1.0.0
jaraco.functools==4.0.0
jaraco.text==3.12.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
parso==0.8.3
path==16.10.0
pexpect==4.9.0
pillow==10.3.0
pip-run==12.6.1
pkginfo==1.10.0
platformdirs==4.2.0
poetry-core==1.9.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.17.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.7.0
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==8.2.3
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
traitlets==5.14.2
trove-classifiers==2024.3.25
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
widgetsnbextension==4.0.10
zipp==3.17.0
